requirements.txt
altair==4.2.2
asn1crypto==1.5.1
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
cryptography==41.0.7
entrypoints==0.4
filelock==3.13.1
gitdb==4.0.11
GitPython==3.1.41
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.3
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.26.3
packaging==23.2
pandas==2.1.4
pillow==10.2.0
platformdirs==3.11.0
protobuf==3.20.3
pyarrow==14.0.2
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
Pympler==1.0.1
pyOpenSSL==23.3.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
six==1.16.0
smmap==5.0.1
snowflake-connector-python==3.6.0
snowflake-snowpark-python==1.9.0
sortedcontainers==2.4.0
streamlit==1.22.0
tenacity==8.2.3
toml==0.10.2
tomlkit==0.12.3
toolz==0.12.0
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
zipp==3.17.0
